Team_14/requirements.txt
Flask==2.3.3
Pillow
numpy
joblib==1.3.2
torch           # keep unpinned or use the proper wheel (see below)
torchvision==0.24.1
timm==0.9.2
facenet-pytorch
opencv-python-headless
tqdm


